+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.1.0
+# 0.1.1
 
 - Added Xrandom and Xrandom64 aliases